enh(ginkgo): unify serial+distributed L1-scaled residual stopping
Merge the serial and distributed L1 stopping criteria into one templated
L1ResidualCriterion<VecType> sharing the norm-factor and residual math. The same
code serves Dense (serial) and distributed::Vector (distributed): for a
distributed vector compute_mean()/compute_norm1() perform the global cross-rank
reductions, so the scaled residual is the global one with no extra MPI code.

- Replace the dormant WIP DistStoppingCriterion with the unified criterion;
  ginkgoL1Stop.cpp is now always built (dist parts under NF_WITH_MPI_SUPPORT)
  and ginkgoL1StopSerial.cpp is removed.
- Add solveWithL1StopDist and wire l1Control through solve_impl_dist + the
  scalar/Vec3 distributed solve paths (mirrors the serial wiring).
- makeConstantLike is the only per-vector-type helper.
